Add PHPUnit, Phake, and a VERY happy test
I decided to take a shot at refactoring the getCacheCode method into a class, as Evan had mentioned in the todo. So far there's no actually useful tests, but I've configured the module for a PSR src directory + added a basic test class template. This should be a good base for extracting and refactoring from here
